[fixed] Grenades are sold in the correct vendor category. [fixed] Items based skills should show up correct in the clients queue, when activated by the items use option. [fixed] Dots don't kill anymore. [fixed] Trap and grenade damage reduce the weapon xp and increase the combat xp (as in r1319).
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1322 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/tangible/terminal/TerminalImplementation.h
@@ -68,7 +68,7 @@ public:
 	static const uint8 MISSION = 4;
 	static const uint8 CLONING = 5;
 	static const uint8 INSURANCE = 6;
-	static const int CAMP = 7;
+	static const uint8 CAMP = 7;
 	static const uint8 STARTINGLOCATION = 8;
 
 	static const uint8 BANK = 42;
